Fixed bugs while debuging
